chorus mgmt api: don't send nil as toBucket in replication responses
If a replication has no Tobucket configured, the name of the source bucket
is used as destination.

Since replications without ToBucket just don't make sense, the mgmt api now
uses the source bucket name as default ToBucket in replication responses.

This can make clients a little bit simpler and more robust, because it is
not necessary to check for nil pointers and to anymore. The client does not
have to know about the meaning of "nil", and act according to the default
anymore.

On the other side, it is not possible to distinguish between "ToBucket not
set" and "ToBucket equals source bucket" anymore.