fix check with custom destination buckets
The check command now determines a replications destination bucket from
the configured replication policy instead of using the source bucket name
as default. This makes the check command usable for replications with
custom destination buckets.

Because this requires the replication policy to be present and
unambiuous, specifying a destination bucket name is now mandatory under
certain conditions. The chorus tests have been adapted accordingly.

"chorctl check" now allows the user to specify the destination bucket
name to use for the check. As a side effect, this makes it possible to
use "check" to compare arbitrary buckets, independently from any configured
replications.