standalone: don't print secrets by default
By default, the standalone proxy prints all configured credentials on
startup, including the secret access keys.
This can be considered a security issue, depending on the environment
where the service is running.

With this commit, the standalone proxy avoids printing the secret keys
unless printing of secrets is explicitly requested in the config.

Signed-off-by: Lukas Thiemeier <thiemeier@strato.de>