introduce ConfAddr type
For addresses in the yaml config files, it is allowed to add "http" or
"https" but not required, as the protocol actually used is determined from
the separate IsSecure setting.

To support that, the code had various places that called TrimPrefix twice
for "http://" and "https://" at runtime. To avoid such unnecessary overhead
the ConfAddr type makes sure the string is stored without the protocol. It
offers various getters to also return the address with its protocol where
needed.

The custom YAML marshal und unmarshal functions are not supported by the
used viper config file framework, which needs some extra code to decode the
new type appropriately. They are still needed, though, because the
standalone's Start() function utilizes a yaml.Marshal / yaml.Unmarshal
combination to implement a deep-copy of the config strucure, among other
places.

Signed-off-by: Jan Schlien <schlien@strato.de>