use own implementation of signS4()
Most of the code was already present to support signature validation. Get
rid of the external dependency on github.com/minio/minio-go/v7/pkg/signer
with the rest of the code added by this commit.

Add "Authorization", "User-Agent" and "Accept-Encoding" to the list of
ignored headers as these are also ignored by the minio-go signer.

Signed-off-by: Jan Schlien <schlien@strato.de>